Service codes organized

+++ lib/data/services/user.service.dart
@@ -1,12 +1,34 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_recipeo/data/models/user_model.dart';
+import 'package:flutter_recipeo/data/services/auth.service.dart';
+import 'package:flutter_recipeo/data/services/storage.service.dart';
+import 'package:flutter_recipeo/locator.dart';
+import 'package:image_picker/image_picker.dart';
 
 class UserService {
   final CollectionReference _usersCollection = FirebaseFirestore.instance.collection('users');
+  final AuthService _authService = locator<AuthService>();
+  final StorageService _storageService = locator<StorageService>();
 
-  Future<void> createUser({required UserModel user}) async {
+  Future<void> createUser({
+    required String displayName,
+    required XFile? image,
+  }) async {
     try {
-      _usersCollection.doc(user.uid).set(user.toJson());
+      final String? uid = _authService.currentUserId;
+      if (uid != null) {
+        String? imageUrl;
+        if (image != null) imageUrl = await _storageService.uploadUserImage(uid: uid, image: image);
+        final UserModel user = UserModel(
+          uid: uid,
+          displayName: displayName,
+          image: imageUrl,
+          posts: [],
+        );
+        await _usersCollection.doc(user.uid).set(user.toJson());
+      } else {
+        throw 'There is no user';
+      }
     } catch (e) {
       rethrow;
     }
@@ -14,7 +36,7 @@ class UserService {
 
   Future<void> updateUser({required UserModel user}) async {
     try {
-      _usersCollection.doc(user.uid).update(user.toJson());
+      await _usersCollection.doc(user.uid).update(user.toJson());
     } catch (e) {
       rethrow;
     }
@@ -30,9 +52,41 @@ class UserService {
     }
   }
 
-  Future<void> deleteUser({required String uid}) async {
+  Future<UserModel> getCurrentUser() async {
+    final String? uid = _authService.currentUserId;
     try {
-      await _usersCollection.doc(uid).delete();
+      if (uid != null) {
+        final DocumentSnapshot snapshot = await _usersCollection.doc(uid).get();
+        final UserModel user = UserModel.fromJson(snapshot.data() as Map<String, dynamic>);
+        return user;
+      } else {
+        throw 'There is no user';
+      }
+    } catch (e) {
+      rethrow;
+    }
+  }
+
+  Future<void> deleteUser({required String uid}) async {
+    final String? uid = _authService.currentUserId;
+    try {
+      if (uid != null) {
+        await _usersCollection.doc(uid).delete();
+      } else {
+        throw 'There is no user';
+      }
+    } catch (e) {
+      rethrow;
+    }
+  }
+
+  Future<void> addPost({required String postId}) async {
+    final String? uid = _authService.currentUserId;
+    try {
+      if (uid != null) {
+        UserModel user = await getCurrentUser();
+        await updateUser(user: user.copyWithNewPost(id: postId));
+      }
     } catch (e) {
       rethrow;
     }
